refactor: Various type / error-handling cleanup
